backend/requirements.txt
click==6.7
Flask==0.12.2
hiredis==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
PyYAML==3.12
redis==2.10.6
Werkzeug==0.12.2
